+++ test/searchfield_test.dart
@@ -34,7 +34,7 @@ void main() {
                   suggestions: ['ABC', 'DEF']
                       .map<SearchFieldListItem<String>>(SearchFieldListItem.new)
                       .toList(),
-                  initialValue: SearchFieldListItem<String>('ABCD'))),
+                  selectedValue: SearchFieldListItem<String>('ABCD'))),
           throwsAssertionError);
 
       await tester.pumpWidget(_boilerplate(
@@ -62,7 +62,7 @@ void main() {
                 width: 1.0)),
         suggestions:
             ['ABC', 'DEF'].map(SearchFieldListItem<String>.new).toList(),
-        initialValue: SearchFieldListItem<String>('ABC'),
+        selectedValue: SearchFieldListItem<String>('ABC'),
       )));
       final finder = find.text('ABC');
       expect(finder, findsOneWidget);
@@ -96,16 +96,24 @@ void main() {
 
     testWidgets('searchfield should show tapped suggestion',
         (WidgetTester tester) async {
-      final controller = TextEditingController();
-      await tester.pumpWidget(_boilerplate(
-          child: SearchField(
-        key: const Key('searchfield'),
-        suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
-            .map(SearchFieldListItem<String>.new)
-            .toList(),
-        controller: controller,
-        suggestionState: Suggestion.expand,
-      )));
+      SearchFieldListItem<String>? selectedValue = null;
+      final boilerPlate = _boilerplate(
+          child: Column(
+        children: [
+          SearchField(
+            key: const Key('searchfield'),
+            suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+                .map(SearchFieldListItem<String>.new)
+                .toList(),
+            selectedValue: selectedValue,
+            onSuggestionTap: (SearchFieldListItem<String> x) {
+              selectedValue = x;
+            },
+          ),
+        ],
+      ));
+      await tester.pumpWidget(boilerPlate);
+      await tester.pumpAndSettle();
       final listFinder = find.byType(ListView);
       final textField = find.byType(TextFormField);
       expect(textField, findsOneWidget);
@@ -114,12 +122,12 @@ void main() {
       await tester.enterText(textField, '');
       await tester.pumpAndSettle();
       expect(listFinder, findsOneWidget);
-      final tapTarget = find.text('ABC');
-      await tester.ensureVisible(tapTarget);
-      expect(tapTarget, findsOneWidget);
-      await tester.tap(tapTarget);
-      await tester.pumpAndSettle(Duration(seconds: 2));
-      expect(find.text('ABC'), findsOneWidget);
+      // tap 2nd item
+      final secondItem = find.text('DEF').first;
+      expect(secondItem, findsOneWidget);
+      await tester.tap(secondItem);
+      await tester.pumpAndSettle();
+      expect(selectedValue!.searchKey, equals('DEF'));
     });
 
     testWidgets('Searchfield should show searched suggestions',
@@ -1141,6 +1149,142 @@ void main() {
     });
   });
 
+  group('Searchfield should respect Suggestion Action', () {
+    testWidgets(
+        'Searcfield should loose focus when suggestion is selected by default',
+        (WidgetTester tester) async {
+      final inputFocus = FocusNode();
+      final searchFocus = FocusNode();
+      SearchFieldListItem<String>? selectedValue = null;
+      final boilerPlate = _boilerplate(
+          child: Column(
+        children: [
+          SearchField(
+            focusNode: searchFocus,
+            key: const Key('searchfield'),
+            suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+                .map(SearchFieldListItem<String>.new)
+                .toList(),
+            selectedValue: selectedValue,
+            onSuggestionTap: (SearchFieldListItem<String> x) {
+              selectedValue = x;
+            },
+          ),
+          TextField(
+            focusNode: inputFocus,
+          )
+        ],
+      ));
+      await tester.pumpWidget(boilerPlate);
+      await tester.pumpAndSettle();
+      final listFinder = find.byType(ListView);
+      final textField = find.byType(TextFormField);
+      expect(textField, findsOneWidget);
+      expect(listFinder, findsNothing);
+      await tester.tap(textField);
+      await tester.enterText(textField, '');
+      await tester.pumpAndSettle();
+      expect(listFinder, findsOneWidget);
+      // tap 2nd item
+      final secondItem = find.text('DEF').first;
+      expect(secondItem, findsOneWidget);
+      await tester.tap(secondItem);
+      await tester.pumpAndSettle();
+      expect(selectedValue!.searchKey, equals('DEF'));
+
+      // check if focus is switched to next field
+      expect(inputFocus.hasFocus, isFalse);
+      expect(searchFocus.hasFocus, isFalse);
+      // show suggestions
+    });
+    testWidgets(
+        'SuggestionAction.next should move focus to next field on selection',
+        (WidgetTester tester) async {
+      final inputFocus = FocusNode();
+      final searchFocus = FocusNode();
+      SearchFieldListItem<String>? selectedValue = null;
+      final boilerPlate = _boilerplate(
+          child: Column(
+        children: [
+          SearchField(
+            focusNode: searchFocus,
+            suggestionAction: SuggestionAction.next,
+            key: const Key('searchfield'),
+            suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+                .map(SearchFieldListItem<String>.new)
+                .toList(),
+            selectedValue: selectedValue,
+            onSuggestionTap: (SearchFieldListItem<String> x) {
+              selectedValue = x;
+            },
+          ),
+          TextField(
+            focusNode: inputFocus,
+          )
+        ],
+      ));
+      await tester.pumpWidget(boilerPlate);
+      await tester.pumpAndSettle();
+      final listFinder = find.byType(ListView);
+      final textField = find.byType(TextFormField);
+      expect(textField, findsOneWidget);
+      expect(listFinder, findsNothing);
+      await tester.tap(textField);
+      await tester.enterText(textField, '');
+      await tester.pumpAndSettle();
+      expect(listFinder, findsOneWidget);
+      // tap 2nd item
+      final secondItem = find.text('DEF').first;
+      expect(secondItem, findsOneWidget);
+      await tester.tap(secondItem);
+      await tester.pumpAndSettle();
+      expect(selectedValue!.searchKey, equals('DEF'));
+
+      // check if focus is switched to next field
+      expect(inputFocus.hasFocus, isTrue);
+      expect(searchFocus.hasFocus, isFalse);
+    });
+
+    testWidgets('Tap Outside should remove focus by default',
+        (WidgetTester tester) async {
+      var outSideTap = false;
+      final focus = FocusNode();
+      final boilerPlate = _boilerplate(
+          child: Column(
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 18.0),
+            child: SearchField(
+              focusNode: focus,
+              key: const Key('searchfield'),
+              suggestions: ['ABC', 'DEF', 'GHI', 'JKL']
+                  .map(SearchFieldListItem<String>.new)
+                  .toList(),
+              onTapOutside: (x) {
+                outSideTap = true;
+              },
+              suggestionState: Suggestion.expand,
+            ),
+          ),
+        ],
+      ));
+      await tester.pumpWidget(boilerPlate);
+      final listFinder = find.byType(ListView);
+      final textField = find.byType(TextFormField);
+      expect(textField, findsOneWidget);
+      expect(listFinder, findsNothing);
+      await tester.tap(textField);
+      expect(focus.hasFocus, isTrue);
+      await tester.enterText(textField, '');
+      await tester.pumpAndSettle();
+      expect(listFinder, findsOneWidget);
+      await tester.tapAt(Offset.zero);
+      await tester.pumpAndSettle();
+      expect(outSideTap, true);
+      expect(focus.hasFocus, isFalse);
+    });
+  });
+
   testWidgets("Test onTapOutside", (widgetTester) async {
     bool outSideTap = false;
     await widgetTester.pumpWidget(_boilerplate(
@@ -1592,7 +1736,7 @@ void main() {
       autoCorrect: false,
       autofocus: true,
       dynamicHeight: true,
-      initialValue: SearchFieldListItem<String>('0'),
+      selectedValue: SearchFieldListItem<String>('0'),
       emptyWidget: Text("Empty"),
       marginColor: Colors.red,
       textInputAction: TextInputAction.done,
@@ -1637,7 +1781,7 @@ void main() {
     expect(searchField.animationDuration, Duration(seconds: 1));
     expect(searchField.autoCorrect, false);
     expect(searchField.dynamicHeight, true);
-    expect(searchField.initialValue, SearchFieldListItem<String>('0'));
+    expect(searchField.selectedValue, SearchFieldListItem<String>('0'));
     expect(searchField.emptyWidget, isA<Text>());
     expect(searchField.marginColor, Colors.red);
     expect(searchField.textInputAction, TextInputAction.done);